pyproject.toml
[project]
name = "nmp-autoavanza"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "pymupdf (>=1.25.5,<2.0.0)",
    "transformers (>=4.51.2,<5.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "progress (>=1.6,<2.0)",
    "ipywidgets (>=8.1.6,<9.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "datasets (>=3.5.0,<4.0.0)",
    "hf-xet (>=1.1.0,<2.0.0)",
    "diffusers (>=0.33.1,<0.34.0)",
    "accelerate (>=1.6.0,<2.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "torchinfo (>=1.8.0,<2.0.0)",
    "torchgan (>=0.1.0,<0.2.0)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "einops (>=0.8.1,<0.9.0)",
    "timm (>=1.0.15,<2.0.0)",
    "sentencepiece (>=0.2.0,<0.3.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "pypdf (>=5.4.0,<6.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "jsonlines (>=4.0.0,<5.0.0)",
    "unstructured[all-docs] (>=0.17.2,<0.18.0)",
    "rapidfuzz (>=3.13.0,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "uvicorn[standard] (>=0.34.3,<0.35.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
